migrate sigma proof pod types
