+++ zk-sdk-pod/src/sigma_proofs.rs
@@ -1,52 +1,43 @@
 //! Plain Old Data types for sigma proofs.
 
-#[cfg(not(target_os = "solana"))]
-use crate::sigma_proofs::{
-    batched_grouped_ciphertext_validity::{
-        BatchedGroupedCiphertext2HandlesValidityProof,
-        BatchedGroupedCiphertext3HandlesValidityProof,
-    },
-    ciphertext_ciphertext_equality::CiphertextCiphertextEqualityProof,
-    ciphertext_commitment_equality::CiphertextCommitmentEqualityProof,
-    grouped_ciphertext_validity::{
-        GroupedCiphertext2HandlesValidityProof, GroupedCiphertext3HandlesValidityProof,
-    },
-    percentage_with_cap::PercentageWithCapProof,
-    pubkey_validity::PubkeyValidityProof,
-    zero_ciphertext::ZeroCiphertextProof,
-};
 use {
-    crate::{
-        pod::{impl_from_bytes, impl_from_str},
-        sigma_proofs::{errors::*, *},
-    },
+    crate::macros::{impl_from_bytes, impl_from_str},
     base64::{prelude::BASE64_STANDARD, Engine},
     bytemuck::{Pod, Zeroable},
     std::fmt,
 };
 
+/// Byte length of a ciphertext-commitment equality proof
+pub const CIPHERTEXT_COMMITMENT_EQUALITY_PROOF_LEN: usize = 192;
+
+/// Byte length of a ciphertext-ciphertext equality proof
+pub const CIPHERTEXT_CIPHERTEXT_EQUALITY_PROOF_LEN: usize = 224;
+
+/// Byte length of a grouped ciphertext for 2 handles validity proof
+pub const GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_LEN: usize = 160;
+
+/// Byte length of a grouped ciphertext for 3 handles validity proof
+pub const GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_LEN: usize = 192;
+
+/// Byte length of a batched grouped ciphertext for 2 handles validity proof
+pub const BATCHED_GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_LEN: usize = 160;
+
+/// Byte length of a batched grouped ciphertext for 3 handles validity proof
+pub const BATCHED_GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_LEN: usize = 192;
+
+/// Byte length of a zero-ciphertext proof
+pub const ZERO_CIPHERTEXT_PROOF_LEN: usize = 96;
+
+/// Byte length of a percentage with cap proof
+pub const PERCENTAGE_WITH_CAP_PROOF_LEN: usize = 256;
+
+/// Byte length of a public key validity proof
+pub const PUBKEY_VALIDITY_PROOF_LEN: usize = 64;
+
 /// The `CiphertextCommitmentEqualityProof` type as a `Pod`.
 #[derive(Clone, Copy)]
 #[repr(transparent)]
-pub struct PodCiphertextCommitmentEqualityProof(
-    pub(crate) [u8; CIPHERTEXT_COMMITMENT_EQUALITY_PROOF_LEN],
-);
-
-#[cfg(not(target_os = "solana"))]
-impl From<CiphertextCommitmentEqualityProof> for PodCiphertextCommitmentEqualityProof {
-    fn from(decoded_proof: CiphertextCommitmentEqualityProof) -> Self {
-        Self(decoded_proof.to_bytes())
-    }
-}
-
-#[cfg(not(target_os = "solana"))]
-impl TryFrom<PodCiphertextCommitmentEqualityProof> for CiphertextCommitmentEqualityProof {
-    type Error = EqualityProofVerificationError;
-
-    fn try_from(pod_proof: PodCiphertextCommitmentEqualityProof) -> Result<Self, Self::Error> {
-        Self::from_bytes(&pod_proof.0)
-    }
-}
+pub struct PodCiphertextCommitmentEqualityProof(pub [u8; CIPHERTEXT_COMMITMENT_EQUALITY_PROOF_LEN]);
 
 const CIPHERTEXT_COMMITMENT_EQUALITY_PROOF_MAX_BASE64_LEN: usize = 256;
 
@@ -70,25 +61,7 @@ impl_from_bytes!(
 /// The `CiphertextCiphertextEqualityProof` type as a `Pod`.
 #[derive(Clone, Copy)]
 #[repr(transparent)]
-pub struct PodCiphertextCiphertextEqualityProof(
-    pub(crate) [u8; CIPHERTEXT_CIPHERTEXT_EQUALITY_PROOF_LEN],
-);
-
-#[cfg(not(target_os = "solana"))]
-impl From<CiphertextCiphertextEqualityProof> for PodCiphertextCiphertextEqualityProof {
-    fn from(decoded_proof: CiphertextCiphertextEqualityProof) -> Self {
-        Self(decoded_proof.to_bytes())
-    }
-}
-
-#[cfg(not(target_os = "solana"))]
-impl TryFrom<PodCiphertextCiphertextEqualityProof> for CiphertextCiphertextEqualityProof {
-    type Error = EqualityProofVerificationError;
-
-    fn try_from(pod_proof: PodCiphertextCiphertextEqualityProof) -> Result<Self, Self::Error> {
-        Self::from_bytes(&pod_proof.0)
-    }
-}
+pub struct PodCiphertextCiphertextEqualityProof(pub [u8; CIPHERTEXT_CIPHERTEXT_EQUALITY_PROOF_LEN]);
 
 const CIPHERTEXT_CIPHERTEXT_EQUALITY_PROOF_MAX_BASE64_LEN: usize = 300;
 
@@ -113,24 +86,8 @@ impl_from_bytes!(
 #[derive(Clone, Copy)]
 #[repr(transparent)]
 pub struct PodGroupedCiphertext2HandlesValidityProof(
-    pub(crate) [u8; GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_LEN],
+    pub [u8; GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_LEN],
 );
-
-#[cfg(not(target_os = "solana"))]
-impl From<GroupedCiphertext2HandlesValidityProof> for PodGroupedCiphertext2HandlesValidityProof {
-    fn from(decoded_proof: GroupedCiphertext2HandlesValidityProof) -> Self {
-        Self(decoded_proof.to_bytes())
-    }
-}
-
-#[cfg(not(target_os = "solana"))]
-impl TryFrom<PodGroupedCiphertext2HandlesValidityProof> for GroupedCiphertext2HandlesValidityProof {
-    type Error = ValidityProofVerificationError;
-
-    fn try_from(pod_proof: PodGroupedCiphertext2HandlesValidityProof) -> Result<Self, Self::Error> {
-        Self::from_bytes(&pod_proof.0)
-    }
-}
 
 const GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_MAX_BASE64_LEN: usize = 216;
 
@@ -155,24 +112,8 @@ impl_from_bytes!(
 #[derive(Clone, Copy)]
 #[repr(transparent)]
 pub struct PodGroupedCiphertext3HandlesValidityProof(
-    pub(crate) [u8; GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_LEN],
+    pub [u8; GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_LEN],
 );
-
-#[cfg(not(target_os = "solana"))]
-impl From<GroupedCiphertext3HandlesValidityProof> for PodGroupedCiphertext3HandlesValidityProof {
-    fn from(decoded_proof: GroupedCiphertext3HandlesValidityProof) -> Self {
-        Self(decoded_proof.to_bytes())
-    }
-}
-
-#[cfg(not(target_os = "solana"))]
-impl TryFrom<PodGroupedCiphertext3HandlesValidityProof> for GroupedCiphertext3HandlesValidityProof {
-    type Error = ValidityProofVerificationError;
-
-    fn try_from(pod_proof: PodGroupedCiphertext3HandlesValidityProof) -> Result<Self, Self::Error> {
-        Self::from_bytes(&pod_proof.0)
-    }
-}
 
 const GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_MAX_BASE64_LEN: usize = 256;
 
@@ -197,30 +138,8 @@ impl_from_bytes!(
 #[derive(Clone, Copy)]
 #[repr(transparent)]
 pub struct PodBatchedGroupedCiphertext2HandlesValidityProof(
-    pub(crate) [u8; BATCHED_GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_LEN],
+    pub [u8; BATCHED_GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_LEN],
 );
-
-#[cfg(not(target_os = "solana"))]
-impl From<BatchedGroupedCiphertext2HandlesValidityProof>
-    for PodBatchedGroupedCiphertext2HandlesValidityProof
-{
-    fn from(decoded_proof: BatchedGroupedCiphertext2HandlesValidityProof) -> Self {
-        Self(decoded_proof.to_bytes())
-    }
-}
-
-#[cfg(not(target_os = "solana"))]
-impl TryFrom<PodBatchedGroupedCiphertext2HandlesValidityProof>
-    for BatchedGroupedCiphertext2HandlesValidityProof
-{
-    type Error = ValidityProofVerificationError;
-
-    fn try_from(
-        pod_proof: PodBatchedGroupedCiphertext2HandlesValidityProof,
-    ) -> Result<Self, Self::Error> {
-        Self::from_bytes(&pod_proof.0)
-    }
-}
 
 const BATCHED_GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_MAX_BASE64_LEN: usize = 216;
 
@@ -245,30 +164,8 @@ impl_from_bytes!(
 #[derive(Clone, Copy)]
 #[repr(transparent)]
 pub struct PodBatchedGroupedCiphertext3HandlesValidityProof(
-    pub(crate) [u8; BATCHED_GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_LEN],
+    pub [u8; BATCHED_GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_LEN],
 );
-
-#[cfg(not(target_os = "solana"))]
-impl From<BatchedGroupedCiphertext3HandlesValidityProof>
-    for PodBatchedGroupedCiphertext3HandlesValidityProof
-{
-    fn from(decoded_proof: BatchedGroupedCiphertext3HandlesValidityProof) -> Self {
-        Self(decoded_proof.to_bytes())
-    }
-}
-
-#[cfg(not(target_os = "solana"))]
-impl TryFrom<PodBatchedGroupedCiphertext3HandlesValidityProof>
-    for BatchedGroupedCiphertext3HandlesValidityProof
-{
-    type Error = ValidityProofVerificationError;
-
-    fn try_from(
-        pod_proof: PodBatchedGroupedCiphertext3HandlesValidityProof,
-    ) -> Result<Self, Self::Error> {
-        Self::from_bytes(&pod_proof.0)
-    }
-}
 
 const BATCHED_GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_MAX_BASE64_LEN: usize = 256;
 
@@ -292,23 +189,7 @@ impl_from_bytes!(
 /// The `ZeroCiphertextProof` type as a `Pod`.
 #[derive(Clone, Copy)]
 #[repr(transparent)]
-pub struct PodZeroCiphertextProof(pub(crate) [u8; ZERO_CIPHERTEXT_PROOF_LEN]);
-
-#[cfg(not(target_os = "solana"))]
-impl From<ZeroCiphertextProof> for PodZeroCiphertextProof {
-    fn from(decoded_proof: ZeroCiphertextProof) -> Self {
-        Self(decoded_proof.to_bytes())
-    }
-}
-
-#[cfg(not(target_os = "solana"))]
-impl TryFrom<PodZeroCiphertextProof> for ZeroCiphertextProof {
-    type Error = ZeroCiphertextProofVerificationError;
-
-    fn try_from(pod_proof: PodZeroCiphertextProof) -> Result<Self, Self::Error> {
-        Self::from_bytes(&pod_proof.0)
-    }
-}
+pub struct PodZeroCiphertextProof(pub [u8; ZERO_CIPHERTEXT_PROOF_LEN]);
 
 const ZERO_CIPHERTEXT_PROOF_MAX_BASE64_LEN: usize = 128;
 
@@ -332,23 +213,7 @@ impl_from_bytes!(
 /// The `PercentageWithCapProof` type as a `Pod`.
 #[derive(Clone, Copy, bytemuck_derive::Pod, bytemuck_derive::Zeroable)]
 #[repr(transparent)]
-pub struct PodPercentageWithCapProof(pub(crate) [u8; PERCENTAGE_WITH_CAP_PROOF_LEN]);
-
-#[cfg(not(target_os = "solana"))]
-impl From<PercentageWithCapProof> for PodPercentageWithCapProof {
-    fn from(decoded_proof: PercentageWithCapProof) -> Self {
-        Self(decoded_proof.to_bytes())
-    }
-}
-
-#[cfg(not(target_os = "solana"))]
-impl TryFrom<PodPercentageWithCapProof> for PercentageWithCapProof {
-    type Error = PercentageWithCapProofVerificationError;
-
-    fn try_from(pod_proof: PodPercentageWithCapProof) -> Result<Self, Self::Error> {
-        Self::from_bytes(&pod_proof.0)
-    }
-}
+pub struct PodPercentageWithCapProof(pub [u8; PERCENTAGE_WITH_CAP_PROOF_LEN]);
 
 const PERCENTAGE_WITH_CAP_PROOF_MAX_BASE64_LEN: usize = 344;
 
@@ -372,23 +237,7 @@ impl_from_bytes!(
 /// The `PubkeyValidityProof` type as a `Pod`.
 #[derive(Clone, Copy, bytemuck_derive::Pod, bytemuck_derive::Zeroable)]
 #[repr(transparent)]
-pub struct PodPubkeyValidityProof(pub(crate) [u8; PUBKEY_VALIDITY_PROOF_LEN]);
-
-#[cfg(not(target_os = "solana"))]
-impl From<PubkeyValidityProof> for PodPubkeyValidityProof {
-    fn from(decoded_proof: PubkeyValidityProof) -> Self {
-        Self(decoded_proof.to_bytes())
-    }
-}
-
-#[cfg(not(target_os = "solana"))]
-impl TryFrom<PodPubkeyValidityProof> for PubkeyValidityProof {
-    type Error = PubkeyValidityProofVerificationError;
-
-    fn try_from(pod_proof: PodPubkeyValidityProof) -> Result<Self, Self::Error> {
-        Self::from_bytes(&pod_proof.0)
-    }
-}
+pub struct PodPubkeyValidityProof(pub [u8; PUBKEY_VALIDITY_PROOF_LEN]);
 
 const PUBKEY_VALIDITY_PROOF_MAX_BASE64_LEN: usize = 88;
