+++ zk-sdk/src/sigma_proofs/mod.rs
@@ -8,7 +8,6 @@
 #![allow(dead_code, unused_imports)]
 
 pub mod errors;
-pub mod pod;
 
 #[cfg(not(target_os = "solana"))]
 pub mod batched_grouped_ciphertext_validity;
@@ -25,32 +24,14 @@ pub mod pubkey_validity;
 #[cfg(not(target_os = "solana"))]
 pub mod zero_ciphertext;
 
-/// Byte length of a ciphertext-commitment equality proof
-pub const CIPHERTEXT_COMMITMENT_EQUALITY_PROOF_LEN: usize = 192;
-
-/// Byte length of a ciphertext-ciphertext equality proof
-pub const CIPHERTEXT_CIPHERTEXT_EQUALITY_PROOF_LEN: usize = 224;
-
-/// Byte length of a grouped ciphertext for 2 handles validity proof
-pub const GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_LEN: usize = 160;
-
-/// Byte length of a grouped ciphertext for 3 handles validity proof
-pub const GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_LEN: usize = 192;
-
-/// Byte length of a batched grouped ciphertext for 2 handles validity proof
-pub const BATCHED_GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_LEN: usize = 160;
-
-/// Byte length of a batched grouped ciphertext for 3 handles validity proof
-pub const BATCHED_GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_LEN: usize = 192;
-
-/// Byte length of a zero-ciphertext proof
-pub const ZERO_CIPHERTEXT_PROOF_LEN: usize = 96;
-
-/// Byte length of a percentage with cap proof
-pub const PERCENTAGE_WITH_CAP_PROOF_LEN: usize = 256;
-
-/// Byte length of a public key validity proof
-pub const PUBKEY_VALIDITY_PROOF_LEN: usize = 64;
+pub use solana_zk_sdk_pod::sigma_proofs::{
+    BATCHED_GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_LEN,
+    BATCHED_GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_LEN,
+    CIPHERTEXT_CIPHERTEXT_EQUALITY_PROOF_LEN, CIPHERTEXT_COMMITMENT_EQUALITY_PROOF_LEN,
+    GROUPED_CIPHERTEXT_2_HANDLES_VALIDITY_PROOF_LEN,
+    GROUPED_CIPHERTEXT_3_HANDLES_VALIDITY_PROOF_LEN, PERCENTAGE_WITH_CAP_PROOF_LEN,
+    PUBKEY_VALIDITY_PROOF_LEN, ZERO_CIPHERTEXT_PROOF_LEN,
+};
 
 #[cfg(not(target_os = "solana"))]
 use {
